Add food size and discount management
- Assign/remove food sizes to/from a food
- Change food availability
- Add, update, delete single or multiple discounts
- Assign/remove discounts to/from one or more foods
- Mark discounts active/inactive
- Added double verification for updates/deletes
- Added UserVerification class
- Moved business logic to services; controllers now only handle requests
- Refactored project structure for readability

+++ src/main/kotlin/com/qinet/feastique/service/AddOnService.kt
@@ -1,36 +1,73 @@
-package com.qinet.feastique.service.addOn
+package com.qinet.feastique.service
 
 import com.qinet.feastique.model.dto.AddOnDto
 import com.qinet.feastique.model.entity.addOn.AddOn
 import com.qinet.feastique.repository.addOn.AddOnRepository
 import com.qinet.feastique.repository.vendor.VendorRepository
 import com.qinet.feastique.security.UserSecurity
+import com.qinet.feastique.security.UserVerification
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
-import java.util.*
 
 @Service
 class AddOnService(
     private val addOnRepository: AddOnRepository,
-    private val vendorRepository: VendorRepository
+    private val vendorRepository: VendorRepository,
+    private val userVerification: UserVerification
 ) {
     @Transactional(readOnly = true)
-    fun getAddOn(addOnId: Long): Optional<AddOn> {
-        return addOnRepository.findById(addOnId)
+    fun getAddOn(
+        id: Long,
+        vendorId: Long,
+        vendorDetails: UserSecurity
+
+    ): AddOn {
+        userVerification.verifyVendor(vendorId, vendorDetails)
+
+        val addOn = addOnRepository.findById(id)
+            .orElseThrow { IllegalArgumentException("No add-on with id: $id.") }
+            .also {
+                if (it.vendor.id != vendorDetails.id) {
+                    throw IllegalArgumentException("You do not have permission to delete add-on: $id")
+                }
+            }
+        return addOn
     }
 
     @Transactional(readOnly = true)
-    fun getAllComplements(vendorId: Long): List<AddOn> {
-        return addOnRepository.findAllByVendorId(vendorId)
+    fun getAllAddOns(
+        vendorId: Long,
+        vendorDetails: UserSecurity
+
+    ): List<AddOn> {
+        userVerification.verifyVendor(vendorId, vendorDetails)
+        val addOns = addOnRepository.findAllByVendorId(vendorId)
+            .takeIf { it.isNotEmpty() }
+            ?: throw IllegalArgumentException("No add-ons found for the vendor $vendorId")
+
+        require(addOns.all { it ->
+            it.vendor.id == vendorDetails.id
+        }) {
+            throw IllegalArgumentException("You (vendor ${vendorDetails.id}) does not have the permission to access these discounts.")
+        }
+        return addOns
     }
 
     @Transactional(readOnly = true)
-    fun getDuplicates(addOnName: String, vendorId: Long): AddOn? {
-        return addOnRepository.findByAddOnNameIgnoreCaseAndVendorId(addOnName, vendorId)
-    }
+    fun getDuplicates(addOnName: String, vendorId: Long): Boolean =
+        addOnRepository.findByAddOnNameIgnoreCaseAndVendorId(addOnName, vendorId) != null
 
     @Transactional
-    fun deleteAddOn(addOn: AddOn) {
+    fun deleteAddOn(
+        id: Long,
+        vendorId: Long,
+        vendorDetails: UserSecurity
+    ) {
+        userVerification.verifyVendor(vendorId, vendorDetails)
+        val addOn = getAddOn(id, vendorId, vendorDetails)
+        if (addOn.vendor.id != vendorDetails.id) {
+            throw IllegalArgumentException("You do not have the permission to delete this add-on.")
+        }
         addOnRepository.delete(addOn)
     }
 
@@ -40,7 +77,7 @@ class AddOnService(
     }
 
     @Transactional
-    fun addAddOn(
+    fun addOrUpdateAddOn(
         addOnDto: AddOnDto,
         vendorDetails: UserSecurity
     ): AddOn {
@@ -56,13 +93,15 @@ class AddOnService(
                     }
                 }
         } else {
-            AddOn()
+            AddOn().apply {
+                this.vendor = vendor
+            }
         }
 
         if(addOnDto.id == null) {
 
             // check if the vendor has already added an add-on with the same name
-            if (getDuplicates(addOnDto.addOnName!!, vendorDetails.id) == null) {
+            if (!getDuplicates(addOnDto.addOnName!!, vendorDetails.id)) {
                 addOn.addOnName = addOnDto.addOnName
             } else {
                 throw IllegalArgumentException("An add-on with the name ${addOnDto.addOnName} already exists. Unable to add a duplicate ")
@@ -72,7 +111,6 @@ class AddOnService(
         }
 
         addOn.price = addOnDto.price ?: throw IllegalArgumentException("Please enter a price.")
-        addOn.vendor = vendor
         addOn = saveAddOn(addOn)
         vendorRepository.save(vendor)
 
@@ -80,4 +118,3 @@ class AddOnService(
 
     }
 }
-
