+++ src/main/kotlin/com/qinet/feastique/service/ComplementService.kt
@@ -1,36 +1,68 @@
-package com.qinet.feastique.service.complement
+package com.qinet.feastique.service
 
 import com.qinet.feastique.model.dto.ComplementDto
 import com.qinet.feastique.model.entity.complement.Complement
 import com.qinet.feastique.repository.complement.ComplementRepository
 import com.qinet.feastique.repository.vendor.VendorRepository
 import com.qinet.feastique.security.UserSecurity
+import com.qinet.feastique.security.UserVerification
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
-import java.util.*
 
 @Service
 class ComplementService(
     private val complementRepository: ComplementRepository,
-    private val vendorRepository: VendorRepository
+    private val vendorRepository: VendorRepository,
+    private val userVerification: UserVerification
 ) {
 
     @Transactional(readOnly = true)
-    fun getComplement(complementId: Long): Optional<Complement> {
-        return complementRepository.findById(complementId)
+    fun getComplement(
+        id: Long,
+        vendorId: Long,
+        vendorDetails: UserSecurity
+
+    ): Complement {
+        userVerification.verifyVendor(vendorId, vendorDetails)
+        val complement = complementRepository.findById(id)
+            .orElseThrow { IllegalArgumentException("No discount found for id: $id") }
+            .also {
+                if (it.vendor.id != vendorDetails.id) {
+                    throw IllegalArgumentException("You do not have permission to delete discount: $id")
+                }
+            }
+        return complement
     }
     @Transactional(readOnly = true)
-    fun getAllComplements(vendorId: Long): List<Complement> {
-        return complementRepository.findAllByVendorId(vendorId)
+    fun getAllComplements(vendorId: Long, vendorDetails: UserSecurity): List<Complement> {
+        val vendor = userVerification.verifyVendor(vendorId, vendorDetails)
+        val complements = complementRepository.findAllByVendorId(vendor.id!!)
+            .takeIf { it.isNotEmpty() }
+            ?: throw IllegalArgumentException("No complements found for the vendor $vendorId")
+        require(complements.all { it ->
+            it.vendor.id == vendorDetails.id
+        }) {
+            throw IllegalArgumentException("Vendor: ${vendorDetails.id}) does not have the permission to access these complements.")
+        }
+        return complements
     }
 
     @Transactional(readOnly = true)
-    fun getDuplicates(complementName: String, vendorId: Long): Complement? {
-        return complementRepository.findByComplementNameIgnoreCaseAndVendorId(complementName, vendorId)
-    }
+    fun getDuplicates(complementName: String, vendorId: Long): Boolean =
+        complementRepository.findByComplementNameIgnoreCaseAndVendorId(complementName, vendorId) != null
+
 
     @Transactional
-    fun deleteComplement(complement: Complement) {
+    fun deleteComplement(
+        id: Long,
+        vendorId: Long,
+        vendorDetails: UserSecurity
+    ) {
+        userVerification.verifyVendor(vendorId, vendorDetails)
+        val complement = getComplement(id, vendorId, vendorDetails)
+        if (complement.vendor.id != vendorDetails.id) {
+            throw IllegalArgumentException("You do not have the permission to delete this complement.")
+        }
         complementRepository.delete(complement)
     }
 
@@ -40,7 +72,7 @@ class ComplementService(
     }
 
     @Transactional
-    fun addComplement(
+    fun addOrUpdateComplement(
         complementDto: ComplementDto,
         vendorDetails: UserSecurity
     ): Complement {
@@ -57,13 +89,15 @@ class ComplementService(
                 }
 
         } else {
-            Complement()
+            Complement().apply {
+                this.vendor = vendor
+            }
         }
 
         if(complementDto.id == null) {
 
             // Check if a vendor has already added a complement with the same name
-            if(getDuplicates(complementDto.complementName!!, vendorDetails.id) == null) {
+            if(!getDuplicates(complementDto.complementName!!, vendorDetails.id)) {
                 complement.complementName = complementDto.complementName ?: throw IllegalArgumentException("Please enter a complement name")
             } else {
                 throw Exception("A complement with the name ${complementDto.complementName} already exist. Unable add a duplicate.")
@@ -73,7 +107,6 @@ class ComplementService(
         }
 
         complement.price = complementDto.price ?: throw IllegalArgumentException("Please enter a price.")
-        complement.vendor = vendor
         complement = saveComplement(complement)
         vendorRepository.save(vendor)
 
@@ -81,4 +114,3 @@ class ComplementService(
     }
 
 }
-
